Updated to move feedback to new class
