+++ src/Dfe.EarlyYearsQualification.Web/Services/Notifications/GovUkNotifyService.cs
@@ -1,3 +1,4 @@
+using Dfe.EarlyYearsQualification.Web.Services.Notifications.Options;
 using Microsoft.Extensions.Options;
 using Notify.Exceptions;
 using Notify.Interfaces;
