+++ src/Dfe.EarlyYearsQualification.Web/Services/Notifications/Options/Feedback.cs
@@ -1,0 +1,8 @@
+namespace Dfe.EarlyYearsQualification.Web.Services.Notifications.Options;
+
+public class Feedback
+{
+    public string TemplateId { get; init; } = string.Empty;
+
+    public string EmailAddress { get; init; } = string.Empty;
+}
