+++ src/Dfe.EarlyYearsQualification.Web/Services/Notifications/Options/NotificationOptions.cs
@@ -1,0 +1,8 @@
+namespace Dfe.EarlyYearsQualification.Web.Services.Notifications.Options;
+
+public class NotificationOptions
+{
+    public string ApiKey { get; init; } = string.Empty;
+
+    public Feedback Feedback { get; init; } = new();
+}
